python_apps/auth_api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auth_api"
version = "0.1.0"
description = "FastAPI-based native authentication API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.2,<2.0.0",
    "sqlalchemy>=1.4.23",
    "alembic>=1.7.4",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.5",
    "bcrypt>=3.2.0",
    "asyncpg>=0.27.0",
    "databases>=0.5.3",
    "httpx>=0.23.0",
    "argon2-cffi>=21.3.0",  # Preferred over bcrypt for password hashing
    "pydantic-settings>=2.0.0",  # For settings management
    "pyotp>=2.8.0",  # For 2FA
    "qrcode>=7.3.1",  # For 2FA QR code generation
    "rate-limiter>=0.0.4",  # For rate limiting
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=22.1.0",
    "isort>=5.10.1",
    "mypy>=0.931",
    "ruff>=0.0.254",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true